\name{STAT600-package}
\alias{STAT600-package}
\alias{STAT600}
\docType{package}
\title{
\packageTitle{STAT600}
}
\description{
\packageDescription{STAT600}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{STAT600}
\packageIndices{STAT600}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{STAT600}

Maintainer: \packageMaintainer{STAT600}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
